+++ webserver/app/containers_api.py
@@ -7,8 +7,8 @@ containers endpoints:
 - POST /registries
 """
 import logging
-from flask import Blueprint, request
 from http import HTTPStatus
+from flask import Blueprint, request
 
 from .helpers.base_model import db
 from .helpers.exceptions import InvalidRequest
@@ -113,7 +113,7 @@ def sync():
         or unintended containers to be used on a node.
     """
     synched = []
-    for registry in Registry.query.filter(Registry.active == True).all():
+    for registry in Registry.query.filter(Registry.active).all():
         for image in registry.fetch_image_list():
             for tag in image["tags"]:
                 if Container.query.filter_by(